CC-1465: rename ucbgUOCApprovalStatus to ucbgUOCbyApprovalStatus
